created separate builders and combined outputs of jupyter code and jupyter
